added role validation login

+++ src/modules/delete_court/app/delete_court_controller.py
@@ -8,6 +8,7 @@ from src.shared.helpers.errors.usecase_errors import NoItemsFound
 from src.shared.helpers.external_interfaces.external_interface import IRequest, IResponse
 from src.shared.helpers.external_interfaces.http_codes import OK, BadRequest, Created, InternalServerError, NotFound
 from src.shared.domain.enums.status_enum import STATUS
+from src.shared.clients.user_api_client import UserAPIClient
 
 class DeleteCourtController:
     
@@ -19,6 +20,10 @@ class DeleteCourtController:
             if request.data.get('number') is None:
                 raise MissingParameters('number')
             
+            if request.data.get('user_from_authorizer') is None:
+                raise MissingParameters('user')
+            
+            user = request.data.get("user_from_authorizer")
             court = self.usecase(number=request.data.get('number'))
             viewmodel = DeleteCourtViewModel(court)
